create hooks.ts and update readme.
